Added Vacation structures and UI.
Added support for vacations as a structured field. The vacations consist of a
From date, a To date, a Period (year) and will have the ability to return how
many workable days happen in that amount of time. This calculation takes only
weekends in consideration, it does not account for national holidays of any
country.

The UI layout is done through the common Collection control link and as such
does not respect the original intention of the mockup. It is, still, a
functional approach and seems acceptable.