Fixed date and timezone offset problem.
Now treating all dates as yyyy-mm-dd strings, to avoid timezone conversions.
When a date object is required, because of controls like datepicker, the date
is transformed in the following way:

    // date format: yyyy-mm-dd
    var year = value.substring(0, 4);
    var month = value.substring(5, 7);
    var day = value.substring(8, 10);

    var date = new Date(year, month - 1, day, 0, 0, 0, 0);

Creating a date object like this gets rid of the timezone conversion, allowing
the new Date object to have the very same date we wanted with the 00:00 hours
of the specific timezone. Tested in Chrome.