Addded new favicon.
Added new favicon and linked it through the master layout file. The used method
for linking the favicon is the most standard one based on the information on
[this Wikipedia article, accesibility section][http://en.wikipedia.org/wiki/Favicon#Accessibility].

It would be a good idea to also allow for other applications to access it
through the root of the web application (on which a new route seems a good idea)
but the cases where this happen are quite seldom and not needed at the moment.