Changed higlight image.
The highlight image has been changed by one located at
http://www.iconfinder.com/icondetails/1564/128/bookmark_favorite_star_icon. This
image was authored by David Vignoni (http://www.icon-king.com/) and the license
for it is LGPL.

As so, the modified image (Unhighlighted.png) will fall onto an open source\
license as well, but not LPGL, but rather the license used in CommonJobs.

The old images (GrayTick and GreenTick) had been deleted.