Fixed Firefox not being able input control keys in numeric fields.
Testing with the salary validation texts, turns out that Firefox does prevent
control keys from executing on the keypress event, as opposed to other browsers
when the `preventDefault()` method is called.

The finished implementation just verifies for those special control keys and
avoids preventing their input. A proper way should be implemented later and a
question in Stack overflow has been posted to check which is the correct way
of implementing this: http://stackoverflow.com/q/10866444/147507