Added new mockups for file search screen.
The file search screen would follow the same guidelines that we have for the
Employee listing and the Applicants listing, except that in the results, instead
of displaying results in a card-listing mode, it would be a linear listing of
documents. The documents should have an icon to display the document type, a
title and a description which would be parsed out of the document.

If possible, it should also display the information relevant to the search that
made the file match a specific search.